feat(FOROME-802): add shadow after horizontal scroll
